+++ ivy/functional/backends/torch/old/general.py
@@ -99,12 +99,6 @@ def identity(n: int, dtype: ivy.Dtype = 'float32', batch_shape: Optional[List[in
         return res
 
 
-def meshgrid(*xs, indexing='ij'):
-    ret = torch.meshgrid(*xs)
-    if indexing == 'xy':
-        # ToDo: verify if this is correct
-        return tuple([torch.transpose(x, 1, 0) for x in ret])
-    return ret
