+++ ivy/functional/ivy/old/general.py
@@ -81,21 +81,6 @@ def maximum(x: Union[ivy.Array, ivy.NativeArray, Number], y: Union[ivy.Array, iv
 
 
 
-def ones(shape: Iterable[int], dtype: Union[ivy.Dtype, str] = 'float32', dev: ivy.Device = None)\
-        -> Union[ivy.Array, ivy.NativeArray]:
-    """
-    Returns a new array of given shape and type, filled with ones.
-
-    :param shape: Shape of the new array, e.g. (2, 3).
-    :type shape: sequence of ints
-    :param dtype: The desired data-type for the array in string format, i.e. 'float32' or 'int64'.
-    Default is 'float32'.
-    :type dtype: data-type string, optional
-    :param dev: device on which to create the array 'cuda:0', 'cuda:1', 'cpu' etc..
-    :type dev: ivy.Device
-    :return: Tensor of ones with the given shape and dtype.
-    """
-    return _cur_framework().ones(shape, dtype, dev)
 
 
 
@@ -155,18 +140,7 @@ def identity(n: int, dtype: ivy.Dtype = 'float32', batch_shape: Iterable[int] = 
     return _cur_framework().identity(n, dtype, batch_shape, dev)
 
 
-def meshgrid(*xs: Iterable[Union[ivy.Array, ivy.NativeArray]], indexing: str = 'ij')\
-        -> Iterable[Union[ivy.Array, ivy.NativeArray]]:
-    """
-    Broadcasts parameters for evaluation on an N-D grid.
 
-    :param xs: input arrays
-    :type xs: sequence of arrays
-    :param indexing: The indexing method, either 'xy' or 'ij'. Default is 'ij'.
-    :type indexing: str, optional
-    :return: list of N-D coordinate arrays for evaluating expressions on an N-D grid
-    """
-    return _cur_framework().meshgrid(*xs, indexing=indexing)
